Changed: Moved all Split/Unsplit Blocks in BC2 Module to Shared Prelude
