+++ projects/dxt-lossless-transform-bc2/src/split_blocks/split/avx512.rs
@@ -420,220 +420,19 @@ pub unsafe fn permute_512_v2(input_ptr: *const u8, output_ptr: *mut u8, len: usi
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::split_blocks::split::tests::{
-        assert_implementation_matches_reference, generate_bc2_test_data,
-        transform_with_reference_implementation,
-    };
-    use core::ptr::copy_nonoverlapping;
-    use dxt_lossless_transform_common::allocate::allocate_align_64;
-    use rstest::rstest;
-
-    type PermuteFn = unsafe fn(*const u8, *mut u8, usize);
-    type PermuteSeparatePointersFn = unsafe fn(*const u8, *mut u64, *mut u32, *mut u32, usize);
+    use crate::test_prelude::*;
 
     #[rstest]
     #[case(permute_512, "avx512_permute")]
     #[case(permute_512_intrinsics, "avx512_permute_intrinsics")]
     #[case(permute_512_v2_intrinsics, "avx512_permute_intrinsics_v2")]
     #[case(permute_512_v2, "avx512_permute_asm_v2")]
-    fn test_avx512_aligned(#[case] permute_fn: PermuteFn, #[case] impl_name: &str) {
-        if !dxt_lossless_transform_common::cpu_detect::has_avx512f() {
+    fn test_avx512_unaligned(#[case] permute_fn: StandardTransformFn, #[case] impl_name: &str) {
+        if !has_avx512f() {
             return;
         }
 
-        for num_blocks in 1..=512 {
-            let mut input = allocate_align_64(num_blocks * 16).unwrap();
-            let mut output_expected = allocate_align_64(input.len()).unwrap();
-            let mut output_test = allocate_align_64(input.len()).unwrap();
-
-            // Fill the input with test data
-            unsafe {
-                copy_nonoverlapping(
-                    generate_bc2_test_data(num_blocks).as_ptr(),
-                    input.as_mut_ptr(),
-                    input.len(),
-                );
-            }
-
-            transform_with_reference_implementation(
-                input.as_slice(),
-                output_expected.as_mut_slice(),
-            );
-
-            output_test.as_mut_slice().fill(0);
-            unsafe {
-                permute_fn(input.as_ptr(), output_test.as_mut_ptr(), input.len());
-            }
-
-            assert_implementation_matches_reference(
-                output_expected.as_slice(),
-                output_test.as_slice(),
-                &format!("{impl_name} (aligned)"),
-                num_blocks,
-            );
-        }
-    }
-
-    #[rstest]
-    #[case(permute_512, "avx512_permute")]
-    #[case(permute_512_intrinsics, "avx512_permute_intrinsics")]
-    #[case(permute_512_v2_intrinsics, "avx512_permute_intrinsics_v2")]
-    #[case(permute_512_v2, "avx512_permute_asm_v2")]
-    fn test_avx512_unaligned(#[case] permute_fn: PermuteFn, #[case] impl_name: &str) {
-        if !dxt_lossless_transform_common::cpu_detect::has_avx512f() {
-            return;
-        }
-
-        for num_blocks in 1..=512 {
-            let input = generate_bc2_test_data(num_blocks);
-
-            // Add 1 extra byte at the beginning to create misaligned buffers
-            let mut input_unaligned = vec![0u8; input.len() + 1];
-            input_unaligned[1..].copy_from_slice(input.as_slice());
-
-            let mut output_expected = vec![0u8; input.len()];
-            let mut output_test = vec![0u8; input.len() + 1];
-
-            transform_with_reference_implementation(input.as_slice(), &mut output_expected);
-
-            output_test.as_mut_slice().fill(0);
-            unsafe {
-                // Use pointers offset by 1 byte to create unaligned access
-                permute_fn(
-                    input_unaligned.as_ptr().add(1),
-                    output_test.as_mut_ptr().add(1),
-                    input.len(),
-                );
-            }
-
-            assert_implementation_matches_reference(
-                output_expected.as_slice(),
-                &output_test[1..],
-                &format!("{impl_name} (unaligned)"),
-                num_blocks,
-            );
-        }
-    }
-
-    #[test]
-    fn avx512_split_blocks_with_separate_pointers_matches_split_blocks() {
-        if !dxt_lossless_transform_common::cpu_detect::has_avx512f() {
-            return;
-        }
-
-        for num_blocks in 1..=512 {
-            let input = generate_bc2_test_data(num_blocks);
-            let len = input.len();
-
-            // Test with the contiguous buffer method
-            let mut output_contiguous = allocate_align_64(len).unwrap();
-
-            // Test with separate pointers
-            let mut alphas = allocate_align_64(len / 2).unwrap(); // 8 bytes per block
-            let mut colors = allocate_align_64(len / 4).unwrap(); // 4 bytes per block
-            let mut indices = allocate_align_64(len / 4).unwrap(); // 4 bytes per block
-
-            unsafe {
-                // Reference: contiguous buffer
-                permute_512(input.as_ptr(), output_contiguous.as_mut_ptr(), len);
-
-                // Test: separate pointers
-                permute_512_with_separate_pointers(
-                    input.as_ptr(),
-                    alphas.as_mut_ptr() as *mut u64,
-                    colors.as_mut_ptr() as *mut u32,
-                    indices.as_mut_ptr() as *mut u32,
-                    len,
-                );
-            }
-
-            // Verify that separate pointer results match contiguous buffer layout
-            let expected_alphas = &output_contiguous.as_slice()[0..len / 2];
-            let expected_colors = &output_contiguous.as_slice()[len / 2..len / 2 + len / 4];
-            let expected_indices = &output_contiguous.as_slice()[len / 2 + len / 4..];
-
-            assert_eq!(
-                alphas.as_slice(),
-                expected_alphas,
-                "AVX512 Alpha section mismatch for {num_blocks} blocks"
-            );
-            assert_eq!(
-                colors.as_slice(),
-                expected_colors,
-                "AVX512 Color section mismatch for {num_blocks} blocks"
-            );
-            assert_eq!(
-                indices.as_slice(),
-                expected_indices,
-                "AVX512 Index section mismatch for {num_blocks} blocks"
-            );
-        }
-    }
-
-    #[rstest]
-    #[case(permute_512_with_separate_pointers, "avx512_permute_separate_pointers")]
-    #[case(
-        permute_512_v2_with_separate_pointers,
-        "avx512_permute_v2_separate_pointers"
-    )]
-    fn test_avx512_separate_pointers_aligned(
-        #[case] permute_fn: PermuteSeparatePointersFn,
-        #[case] impl_name: &str,
-    ) {
-        if !dxt_lossless_transform_common::cpu_detect::has_avx512f() {
-            return;
-        }
-
-        for num_blocks in 1..=512 {
-            let input = generate_bc2_test_data(num_blocks);
-            let len = input.len();
-
-            // Reference: contiguous buffer
-            let mut output_expected = allocate_align_64(len).unwrap();
-            transform_with_reference_implementation(
-                input.as_slice(),
-                output_expected.as_mut_slice(),
-            );
-
-            // Test: separate pointers
-            let mut alphas = allocate_align_64(len / 2).unwrap(); // 8 bytes per block
-            let mut colors = allocate_align_64(len / 4).unwrap(); // 4 bytes per block
-            let mut indices = allocate_align_64(len / 4).unwrap(); // 4 bytes per block
-
-            alphas.as_mut_slice().fill(0);
-            colors.as_mut_slice().fill(0);
-            indices.as_mut_slice().fill(0);
-
-            unsafe {
-                permute_fn(
-                    input.as_ptr(),
-                    alphas.as_mut_ptr() as *mut u64,
-                    colors.as_mut_ptr() as *mut u32,
-                    indices.as_mut_ptr() as *mut u32,
-                    len,
-                );
-            }
-
-            // Verify that separate pointer results match contiguous buffer layout
-            let expected_alphas = &output_expected.as_slice()[0..len / 2];
-            let expected_colors = &output_expected.as_slice()[len / 2..len / 2 + len / 4];
-            let expected_indices = &output_expected.as_slice()[len / 2 + len / 4..];
-
-            assert_eq!(
-                alphas.as_slice(),
-                expected_alphas,
-                "{impl_name}: Alpha section mismatch for {num_blocks} blocks"
-            );
-            assert_eq!(
-                colors.as_slice(),
-                expected_colors,
-                "{impl_name}: Color section mismatch for {num_blocks} blocks"
-            );
-            assert_eq!(
-                indices.as_slice(),
-                expected_indices,
-                "{impl_name}: Index section mismatch for {num_blocks} blocks"
-            );
-        }
+        // AVX512 implementation processes 256 bytes per iteration, so max_blocks = 256 * 2 / 16 = 32
+        run_standard_transform_unaligned_test(permute_fn, 32, impl_name);
     }
 }
