+++ projects/dxt-lossless-transform-bc2/src/split_blocks/unsplit/avx2.rs
@@ -385,89 +385,16 @@ pub unsafe fn avx2_shuffle_with_components(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::split_blocks::split::tests::generate_bc2_test_data;
-    use crate::split_blocks::split::u32;
-    use crate::split_blocks::unsplit::tests::assert_implementation_matches_reference;
-    use dxt_lossless_transform_common::allocate::allocate_align_64;
-    use rstest::rstest;
-
-    type DetransformFn = unsafe fn(*const u8, *mut u8, usize);
+    use crate::test_prelude::*;
 
     #[rstest]
     #[case::avx2_shuffle(avx2_shuffle, "avx2_shuffle")]
-    fn test_avx2_aligned(#[case] detransform_fn: DetransformFn, #[case] impl_name: &str) {
-        if !dxt_lossless_transform_common::cpu_detect::has_avx2() {
+    fn test_avx2_unaligned(#[case] detransform_fn: StandardTransformFn, #[case] impl_name: &str) {
+        if !has_avx2() {
             return;
         }
 
-        // Test with different block counts to ensure they all work correctly
-        for block_count in 1..=512 {
-            // Generate test data
-            let original = generate_bc2_test_data(block_count);
-            let mut transformed = allocate_align_64(original.len()).unwrap();
-            let mut reconstructed = allocate_align_64(original.len()).unwrap();
-
-            unsafe {
-                // Transform the original test data
-                u32(original.as_ptr(), transformed.as_mut_ptr(), original.len());
-
-                // Re-transform it back using the implementation under test
-                (detransform_fn)(
-                    transformed.as_ptr(),
-                    reconstructed.as_mut_ptr(),
-                    transformed.len(),
-                );
-            }
-
-            // Verify the results match
-            assert_implementation_matches_reference(
-                original.as_slice(),
-                reconstructed.as_slice(),
-                impl_name,
-                block_count,
-            );
-        }
-    }
-
-    #[rstest]
-    #[case::avx2_shuffle(avx2_shuffle, "avx2_shuffle")]
-    fn test_avx2_unaligned(#[case] detransform_fn: DetransformFn, #[case] impl_name: &str) {
-        if !dxt_lossless_transform_common::cpu_detect::has_avx2() {
-            return;
-        }
-
-        // Test with different block counts to ensure they all work correctly
-        for block_count in 1..=512 {
-            // Generate test data
-            let original = generate_bc2_test_data(block_count);
-
-            // Create unaligned buffers (allocate an extra byte and offset by 1)
-            let mut unaligned_transformed = vec![0u8; original.len() + 1];
-            let mut unaligned_reconstructed = vec![0u8; original.len() + 1];
-
-            unsafe {
-                // Transform the original test data
-                u32(
-                    original.as_ptr(),
-                    unaligned_transformed.as_mut_ptr().add(1),
-                    original.len(),
-                );
-
-                // Re-transform it back using the implementation under test
-                (detransform_fn)(
-                    unaligned_transformed.as_mut_ptr().add(1),
-                    unaligned_reconstructed.as_mut_ptr().add(1),
-                    unaligned_transformed.len() - 1,
-                );
-            }
-
-            // Verify the results match
-            assert_implementation_matches_reference(
-                original.as_slice(),
-                &unaligned_reconstructed[1..],
-                impl_name,
-                block_count,
-            );
-        }
+        // AVX2 implementation processes 128 bytes per iteration, so max_blocks = 128 * 2 / 16 = 16
+        run_standard_untransform_unaligned_test(detransform_fn, 16, impl_name);
     }
 }
